requirements.txt
numpy==1.24.3
tensorflow==2.15.0
opencv-python==4.8.0.74
matplotlib==3.7.2
seaborn==0.12.2
scikit-learn==1.3.0
joblib==1.3.2
flask==2.3.3
pillow==10.0.0
flask-cors==4.0.0
